Still working on getting already added A* to work with dinos, also adding new state to transition between eating
The A* needs to be looked at as when used on the dinos with an added food source they will not travel to the source at the moment.
Also having grazing state check if there is food nearby and if food is found it then transitions to travelling towards that food source. This is being done to reduce the complexity of the grazing state where this would all have been done in one state