Almost fixed align function in fsm
Align has not been fucntion properly in the state machine due to it requiring the object it is facing to have the agent class. However adding and enabling this class causes food sources to move which is undesired. New "agent" script may need to be made for food sources to allow them to be faced towards without having them move