Basic Eating State Complete
Multiple Ankys can now correctly transition from the eating state. This can be a bit slow which is likely due to the animator state machine not being optimised for being used as a general state machine in this way. Anky still won't face towards food and does not yet use an A* algorithm but this way at least creates a functional state.