Beginning Final States for Anky
Beginning work on the last few states for the anky. This state machine will end up being very similar to the one for Rapty as rapty should only need small changes to use ankys as food. States for when anky sees and runs from raptys just need to be finished