Final Push
Overall, there is just so much more that I hoped to achieve, but time constraints limited me to what I could achive in the timeframe.

Note: If you wish to see the different scripts in action you'll need to change the target in Rapty's Face script.  Water Source shows off what I've done for drinking, and a dead anky to see it eat. If you want to see it die, just give it time.